weird type change

+++ src/cpu.cpp
@@ -132,13 +132,13 @@ void CPU::L8U() {
 void CPU::J() { program_counter = 4 * instruction_context.immediate; }
 
 void CPU::S16() {
-    uint32_t effective_address =
+    uint16_t effective_address =
         registers[instruction_context.reg_a] + instruction_context.immediate;
     this->console->memory.w16u(effective_address, instruction_context.reg_b);
 }
 
 void CPU::S8() {
-    uint32_t effective_address =
+    uint16_t effective_address =
         registers[instruction_context.reg_a] + instruction_context.immediate;
     this->console->memory.w8u(
         effective_address,
